requirements.txt
azure-common==1.1.14
azure-mgmt-nspkg==2.0.0
azure-mgmt-resource==2.0.0
azure-mgmt-subscription==0.2.0
azure-nspkg==2.0.0
azure-storage==0.36.0
azure-servicebus==0.21.0
azure-functions==1.0.0a3
azure-cosmosdb-table==1.0.3
mock==2.0.0
click==6.7